Fix a bug in parented_rewrite_pre, and make traversal in order
Summary:
Parented_rewrite_pre was reversing children when recreating the node, which caused the output of changed nodes to be backwards (a changed xhp_expression came back as xhp_close, body, open instead of open, body, close)

Also changed this to be in order traversal of the node (instead of folding right on children). This has the nice property of being able to perform a transformation on the first instance of some issue by combining the closure with a reference.

Differential Revision: D4594098

fbshipit-source-id: 627ab70990b10babafdcd4f6bc1a386010417f95