Cut-over of `hh_emitter` to use the Full Fidelity Parser
Summary: Does what it says on the tin. Adding hubyrod as a reviewer, because he knows more about the time-line towards killing `hh_emitter`. If this diff doesn't get in the way of anything and looks good, I would still like to push this through to clean up the legacy-calls.

Reviewed By: hubyrod

Differential Revision: D4620273

fbshipit-source-id: 80d2cb98e0f73339f5c8cfc8c3834b123d65d82a