Make copy_ptr explicitly copy-on-write
Summary:
Rather than copying on construction, make it expose a read-only
smart-pointer interface by default, and add a mutate() method to
return a mutable copy.

Make the underlying implementation std::shared_ptr, so that we can
avoid copying when the ref-count is one, and add an emplace method to
make the api more agnostic to the implementation details.

Reviewed By: swtaarrs, mxw

Differential Revision: D4628131

fbshipit-source-id: 3c0a7bef5f57fc384869edcb110b2d9b51e371b5