Implement finally-blocked continue and break
Summary:
Finally-blocked continue and break are now implemented for the scenario where there is exactly one continue or break in the try-finally block.

The case where there are two or more is not yet implemented. That will come in the next diff.

The code should work for PHP-style multi-level continue and break, but we're not set up to test that yet.  I'll work on that later.

As expected, we've changed a bunch of the labels because fault and catch labels are not yet canonicalized.

Reviewed By: oulgen

Differential Revision: D4639108

fbshipit-source-id: 8d643b4d3b1933d1d70bd0da1d25c2b36179e20e