Fix frame-state tracking for ifThenElse() with an unreachable else
Summary:
If one uses ifThenElse() and the else case ends up being unreachable,
then we'll end up using the then case's frame-state in the else case. This is
because else won't have any predecessors. This causes issues because the then
case may manipulate the stack. Fix this by explicitly specifying that the else
case's predecessor is the branch case.

Reviewed By: mxw

Differential Revision: D4509883

fbshipit-source-id: 5d11d4f49b8f3213af304a965936b78d4fab2f81