Add basic support for class-ref slot immediates in HHBC
Summary:
Add basic support for class-ref slot immediates in HHBC. Class-ref slots are
effectively scratch registers where class-refs (flavor A) can be read from and
written to. Unlike values on the stack, class-ref slots are always explicitly
addressed by the bytecode op. In this manner they are similar to iterator slots.

Class-ref slots will allow us to never have to manipulate class-refs on the
normal eval stack, which will allow us to get rid of KindOfClass. Class-refs are
somewhat special in that they're their own flavor and can only be on the stack
(never locals or the heap).

This commit adds two new immediate types (CAR and CAW) which are encoded the
same as IVA, but represent a class-ref slot being read from or written to. The
advantage of using two different immediates is that it allows us to
automatically determine whether an instruction reads from or writes to a
slot. Add all the usual plumbing for the new immediate types, including in
HHBBC.

For simplicity, the emitter will automatically assign class-ref slots to the
instructions which require them. Since class-refs are currently managed on the
stack with the rest of the values, allocating class-ref slots is easy. Whenever
they would normally be pushed onto the eval stack, a new slot is assigned, and
when it would be pushed off, that slot is freed. Since slots are always assigned
in-order, this means an integer is sufficient to track the class-ref slot
state. Since the emitter is assigning class-ref slots, provide a "placeholder"
value which is provided in place of the class-ref immediate, which will be
replaced with an actual slot by the emitter.

The emitter also tracks the maximum number of class-ref slots alive in the
function, which is stored in the function's metadata. To prevent making Func
larger, we normally store the number of class-ref slots in just two bits, which
allows for up to 3 class-ref slots. Needing more than one class-ref slot is very
rare and needing more than two essentially doesn't happen. If we happen to need
more than 3, we use ExtendedSharedData instead to store the full count.

Note that this diff only add supports for the class-ref slots. Nothing at
runtime actually uses them. The actual class-refs are still pushed/popped onto
the stack as usual.

Reviewed By: markw65

Differential Revision: D4465986

fbshipit-source-id: 37318711358d72f4b173da61dc0ffa91dbb22397