Trace changes to pre-live ActRec in frame-state
Summary:
In certain situations during a function call sequence, we may not know
the function (or its context), so we write null pointers to the pre-live ActRec
slots. We then execute a HHIR operation which does the proper lookup, and it
writes the results to the pre-live ActRec. The issue is that frame-state wasn't
handling these operations, so frame-state continued to think the ActRec
contained the null pointers. This is usually harmless, but it can cause issues
if we decide the inline the function we're calling. We'll think the context is a
null pointer even if it shouldn't be. Note that this ituation cannot usually
happen (not knowing the function/context during the SpillFrame, but inlining the
call). This was noticed when switching to use class-ref slots. Until
memory-effects is done, we forget the type of ClassData's we write to the frame,
which can trigger the above.

Fix this by processing the HHIR ops which write to the pre-live ActRec. Since we
(usually) cannot predict what it will write to the ActRec, simply assume the
context is now a TCtx. Add a marker to prevent the inliner from attempting to
inline the call corresponding to this ActRec, as the inliner requires more
detailed context information than we can provide. This won't cause a regression
because (as said above) we can't usually inline a call when we don't know the
context anyways.

While I was at it, I noticed the inliner attempts to load the context using a
LdStk on a location corresponding to the pre-live ActRec. This has two
issues. The first is that you cannot load a TCls from the pre-live ActRec,
because we store TCctx in them instead. The second is that LdStk is meant to
load TypedValues and the context isn't stored as a TypedValue within the
ActRec. Instead invent a new HHIR opcode LdARCtx, which is specialized to load
the object or class context out of a pre-live ActRec.

Reviewed By: markw65

Differential Revision: D4484907

fbshipit-source-id: 51d16e635a6a5a1d0f97cddda552b69d7c5644d7