Actually use class-ref slots at runtime
Summary:
Actually use the class-ref slots at runtime. Both HHBBC, the interpreter, and
the JIT treat the class-ref slots as authoritative. We still continue to push
and pop values on the stack. This is because we want to minimize changes and
don't want to change the depth of the stack from normal. To catch errors (where
we're still using a class-ref from the stack), push InitNull where-ever we would
have pushed a class-ref. The actual push/popping will be removed later.

Class-ref slots have initialization semantics somewhat similar to iterator
slots. They start out as uninitialized, and are initialized by writing to
them. It is illegal to write to an already initialized class-ref slot or to read
from an uninitialized class-ref slot. Reading from an initialized class-ref slot
makes it become uninitialized as a side-effect. This may sound odd, but it
actually mirrors the previous stack semantics. In debug builds, we may trash a
class-ref slot with a poison value when it becomes uninitialized.

Add three new HHIR ops. LdClsRef, StClsRef, and KillClsRef. LdClsRef and
StClsRef have obvious meaning, but KillClsRef explicitly represents a class-ref
slot becoming uninitialized. This is useful for memory-effects because it
indicates that the value in the slot no longer needs to be kept. LdClsRef is
just a pure load. KillClsRef usually compiles to no code, but in debugging
builds it may poison the class-ref slot.

Class-ref slots are allocated in the frame after the iterator slots. Since they
only occupy 8 bytes, we can pack two of them into the normal space for a Cell.

Add a m_magic entry to Class, just like we have for Func. This makes it easier
to determine if we've gotten a garbage value out of a class-ref slot (in
debugging builds).

Note that this diff on its own is likely to be a regression. The reason is
because we do not have memory-effects or frame-state support for these new
instructions, so we'll read/write unnecessarily. This will be introduced in a
later diff.

Reviewed By: swtaarrs

Differential Revision: D4471378

fbshipit-source-id: 8c9bfb5eeeb04aeea7e03ab1d05c5116ee67bd28