Remove KindOfClass and TStkElem
Summary:
This is the payoff. Remove KindOfClass from Datatype and TStkElem from the JIT
type-system. Now that class-refs can only be stored in class-ref slots, there's
no need for a value in Datatype to represent them. TStkElem represented the
union of TGen and TCls and was exclusively used for types that can be present on
the eval stack. Since TCls can never appear on the eval stack now, its not
needed, so replace all references to just TGen. Likewise, since Class* can't
appear in a typed-value now, remove the pcls from it.

Reviewed By: markw65

Differential Revision: D4510875

fbshipit-source-id: 4c6edc7ceb18143d5abf4780e7d724e5860500b6