Add class-ref slot support to verifier
Summary:
Add support in the verifier to check all the class-ref slot invariants in
HHBC. While we're at it, add a few other checks for other opcodes. Namely, the
ops which index into the stack are verified that they actually index to a valid
stack element. I ran into this when converting over to use class-ref slots and
was surprised the verifier didn't flag this.

Reviewed By: swtaarrs

Differential Revision: D4511104

fbshipit-source-id: dcf44c0375c89f1eb97ad8272e52c0fce9ff7f93