Implement class-ref DCE in HHBBC
Summary:
Implement DCE of class-ref slot usages in HHBBC. This will remove individual
usages (in certain circumstances) of class-ref slots, and if a class-ref slot
becomes completely unused, it will be eliminated entirely from the func. Only a
very limited number of instructions are supported (the same as when class-refs
were on the stack). Like the elimination of stack values, the algorithm mainly
only works within basic blocks.

To remove ambiguity, rename the existing data structures which refer to locals
to have local in their name, since the class-ref slots have similar data
structures).

Reviewed By: markw65

Differential Revision: D4619623

fbshipit-source-id: b92427b4b3049d0396ec9330b811d41f2a57f224