Allow for (P)DCE of DefInlineFP with class-ref slots
Summary:
Previously, when class-ref slots were stored on the stack, there was no issue
with (partial) DCE of DefInlineFP, as their loads and stores were always indexed
off the stack pointer. Now class-ref slots are stored in a manner similar to
iterators and locals, and thus are indexed off the frame pointer. However,
unlike locals, it is not possible to rewrite the loads and stores to them to be
stack pointer relative. This is for two reasons: The first is that class-refs
can never be stored in the stack now, and thus it would be illegal to use a
LdStk/StStk to manipulate them. The second is that we no longer use TypedValues
for class-refs, thus LdStk/StStk wouldn't be able to load/store them anyways.

Its not actually necessary to rewrite the class-ref loads and stores to be
relative to the stack pointer however. Instead, we can rewrite them to be
indexed off of the caller's frame pointer (or even higher as necessary). Since
the actual memory address that is accessed is just the frame pointer plus some
offset calculated from the slot number, we can adjust the slot number so that
using a different frame pointer will result in the same address.

In principle, the same technique can be used with iterators as well.

Reviewed By: paulbiss

Differential Revision: D4619653

fbshipit-source-id: cbb92046b58015e316ae9d2613d89eef1c467a84