Add SuppressHackArrayCompatNotices RAII class; use in SourceRootInfo
Summary:
SourceRootInfo uses += internally, which was triggering Hack array
compat notices. However, this is a legitimate usage, as these are not candidates
to be converted to Hack arrays.

Add a new class called SuppressHackArrayCompatNotices; This class is meant to be
used in a RAII manner and will suppress the emission of Hack array compat
notices when it is in scope. Use this in SourceRootInfo.

Reviewed By: mxw

Differential Revision: D4619148

fbshipit-source-id: 8efb509564e49bc66b12fbb0dcb4997adefbb796