Implement locals as tagged data, not as integers and strings
Summary:
We have a tag for whether a variable is an numbered or named local but we were not using it. We now treat locals polymorphically as locals, not as strings or integers.

I've also made an API in the label generator for unnamed local generation. I'll change it so that it is independent of label in a later diff.

Reviewed By: oulgen

Differential Revision: D4643615

fbshipit-source-id: e7ab2f79960cad72b076abfe44d0e5c9803eb3a1