Allow static arrays to skip low memory
Summary:
Nothing requires them to be in low memory, even in LOWPTR builds - and
hhbbc can create a lot of static arrays while optimizing; enough that
some static strings don't fit in low memory.

Add an option to enable the use of low memory, default it to true, and
disable it for hhbbc.

Reviewed By: swtaarrs, binliu19

Differential Revision: D4642725

fbshipit-source-id: 57c2fd396b3af10c40255636e60fa427d8f3a43c