Capture tpenv at time of add_todo constraint check
Summary:
Consider the following example:

  function foo<T as arraykey>(T $x): void { }
  class C<+T> { }
  class D extends C<arraykey> { }
  function test<Tu>(C<Tu> $x, Tu $y): void {
    if ($x instanceof D) {
      // Here we know Tu <: arraykey but the constraint is checked later
      foo($y);
    }
  }

Currently the Typing_env.add_todo function registers a function to be run
potentially at the end of checking `test`, that checks that the constraint
`T as arraykey` on `foo` is satisfied by the actual type argument `Tu`. But
by this point the type parameter environment `tpenv` will no longer have the
information that `Tu <: arraykey` because the refinement on `$x` is no
longer in scope.

The solution is to capture `tpenv` at the point that `add_todo` is called, and
temporarily add this to the environment under which the registered function is
called.

Reviewed By: dlreeves

Differential Revision: D4645478

fbshipit-source-id: dbc1eeed41d2cb4c90e88e34cea58f158ecd45bc