Separate iterator ids from label ids
Summary:
We need to generate unique iterator ids; we were using the label generator, but that's wrong.

The approach now is still wrong; we actually need to *nest* the ids so that

    foreach(...) { foreach (...) { ... } }
    foreach(...) { ... }

generates id zero for the two outer iterators and id one for the inner.  But that will happen in a later diff.

Right now I just want to (1) separate iterator id generation from label id generation, and (2) wrap up the int in a wrapper type so that we never accidentally use it as a label, local, etc.

Reviewed By: oulgen

Differential Revision: D4653303

fbshipit-source-id: b06c37dd9a119a62d0948a85145c67c6645c43c6