Continue to reorganize label code
Summary:
A label was a pair of (int, flavor). But this is a weird way to represent a tagged integer in OCaml. I've reworked the code so that the kind is the constructor for the label.

We are still however passing around labels as plain ints. In the next diff I'll clean this up further.

Reviewed By: oulgen

Differential Revision: D4653733

fbshipit-source-id: caa19b1e58bafa93a6c38b474768320b078369df