Fix a bug in MixedArray::CopyMixed
Summary:
When using CopyMixed to convert a MixedArray to a Dict, we would throw
if we found a true reference element. In debug builds, we could fail
an assert in checkInvariants if m_pos was past the reference, so set
m_pos to zero before destroying the newly created array.

While here, I noticed that throwRefInvalidArrayValueException only
cares about the HeaderKind of the array its passed (which is just used
for the error message); so avoid the ScopeGuard by destroying ad
first, and passing in staticEmptyDictArray().

Reviewed By: ricklavoie

Differential Revision: D4657301

fbshipit-source-id: e2a828c5f1b37c851e787293177179cec98f9e62