Generate memoized wrapper function
Summary:
A first cut at memoization. We implement memoization by renaming the original function, and then generating a function with the original function's name. That new function then builds a dictionary mapping inputs to outputs.

So far we rename the original method, and generate a mapper function that does nothing.

Reviewed By: oulgen

Differential Revision: D4662957

fbshipit-source-id: 422438a90b1791406bd8d83968f6f7cb64dd5a9b