Resolve constants in a pre-pass
Summary:
The normal analyze_iteratively pass analyzes all methods/classes, then
updates the index, figures out which methods/classes might need
re-analyzing, and repeats.

If there are a lot of constants and class constants that haven't been
resolved yet, we'll optimize a lot of methods without knowing their
values; worse, if the constants are defined in terms of other
constants they won't be resolved on the second pass either (and we'll
have pointlessly re-optimized all the methods that depend on them).

This adds a constant resolving pass that only includes functions that
define constants, and 86cinit methods. Once thats done, it throws in
any methods that use AddElemC or AddNewElemC, and optimizes them all
with the current state of the index. This will convert many of those
arrays to static arrays, and prevent repeatedly analyzing the long
sequences of AddElems in later analyze passes.

Reviewed By: mxw

Differential Revision: D4630112

fbshipit-source-id: becb21282f3a95f7ad1521d3db49b16bfe53f737