Remove HackArrCompatNotice for Idx with null key
Summary:
Passing null as the key to Idx always returns the default value. This
is *not* an implicit key conversion (it returns the default even if the empty
string or 0 is present as a key), and this behavior doesn't change with Hack
arrays. Therefore we shouldn't be raising a HackArrCompatNotice for it.

Reviewed By: swtaarrs

Differential Revision: D4665265

fbshipit-source-id: 31c04015347bec725aa9d0ba2521f4005c133661