Cut-over of `hh_single_compile` to use the Full Fidelity Parser
Summary: Moving `hh_single_compile` to use the Full Fidelity Parser so as to get out of the critical path for codegen. A test-case for the codegen (`test_binops.php`) exposed a missing case in `pBinOp`. Should that fix be different diff? Yes, if you're very puritanical about it. I'm not.

Reviewed By: ericlippert

Differential Revision: D4643315

fbshipit-source-id: 59f12098053bbf1140680d57a253c04298695e5f