Finally fixing the string madness
Summary:
There was much more work in this diff then the code suggests. Review should be thorough, so I'm only adding ericlippert and eshrews, for them having the context on whitespace in XHP land.

I have been struggling with restoring semantically significant whitespace in XHP expressions. I ended up with this rather brute-force solution. Curious as to whether this would break anything subtle. Here's my thinking:
 1. Spacing in XHP is relevant //upto// one space, i.e. any sequence of whitespace characters get substituted by a single space character.
 2. It is only relevant in "raw" XHP bodies, i.e. not inside elements, attributes and/or nested braced expressions.
 3. Having a consistent approach to trivia (stick it in trailing trivia of the previous token), but considering that this is the only (to my knowledge) place where whitespace has semantics, it's worth breaking convention and reasonable to ask users to understand this.

> From these three points, I concluded that it would be okay to take the default approach in XHP to make //all// trivia *leading*, except for `XHPBody` tokens.

The beauty is that this change already correctly propagates through the lowerer where `s/\s+/ /g` normalises the nested whitespace (without breaking the beautiful FFP approach of not creating individual strings for bits of the content, but pointing to the original blob).

Reviewed By: ericlippert

Differential Revision: D4634979

fbshipit-source-id: 8743041e67a41bc0a786da7c13b01e0e005be3cf