Drop support for instrumentation-based profile counters
Summary:
These were useful in the days of tracelets and before we had support for dumping
region descriptors (which include the region block profile counters).  The
original counters that we used to emit in the JIT to populate these data
("translation counters") have even been removed by now.  We're currently trying
to populate these with the profile counters, which are only (sometimes)
available for profile translations.  This is not very useful, so just get rid of
this support and clean up the code.

Reviewed By: paulbiss

Differential Revision: D4674288

fbshipit-source-id: ac43f75fd6d5b2017987d1ef35bdc251a739e703