Throw an exception when Redis returns an OOM error or aborts a transaction
Summary:
When Redis is out of memory, the hhvm Redis client blocks until the socket times out while executing a transaction.

The problem is that `exec` calls `flushCallbacks` which ignores the errors coming back from Redis and is stuck waiting to get the responses for the commands in the transaction, which are never going to come.

More specifically, when Redis is OOM:
1. When the client queues a command in a transaction, Redis returns - `OOM command not allowed when used memory > 'maxmemory'`.
2. When the client tries to execute this transaction, Redis returns - `Transaction discarded because of previous errors`.

In this PR, both of these errors are thrown as exceptions so the caller knows that there is a problem with Redis.

1. I have used Redis errors as is when throwing exceptions. `Transaction discarded because of previous errors` does not look very intuitive as a standalone exception. Is there is a better way here?

1. I could not find a test file for existing exception c
Closes https://github.com/facebook/hhvm/pull/7644

Reviewed By: aorenste, Orvid

Differential Revision: D4508262

Pulled By: mofarrell

fbshipit-source-id: 0092d0b647386e17c254d3e0626675b138c2893b