Track local statics
Summary:
We don't track the types of boxed locals, because almost anything
could change them. But local statics are special; unless we explicitly
box it again after the StaticLocInit, nothing else can get to the
reference.

This tracks locals which are bound to statics, and the types of the
statics, and uses that to infer better types for the locals.

In particular, this means we can often infer return types, or even
return values for functions that return static locals.

Reviewed By: ricklavoie

Differential Revision: D4652931

fbshipit-source-id: 14199084587b2661518df8f6f5f350c3e7d3c1b1