Optimize no-op conditional jumps
Summary:
If a conditional branch branches around a block, and we optimize away
the contents of the block, we end up with a branch around a block
containing a single nop (hhbbc doesn't allow empty blocks).

This follows "nop" blocks at the fallthrough and target, and if they
end up at the same place, eliminates the jmp.

Reviewed By: ricklavoie

Differential Revision: D4689119

fbshipit-source-id: 5df824a31b597513859278ee17eabc5a6fe96928