Optimize CheckProp
Summary:
I noticed that with constant prop disabled in hphpc, we still create a
lot more 86pinits than before. Looking at them, it turns out that
we've eliminated the InitProps, but left the CheckProp/JmpNZ guards.

After D4689119, the JmpNZs are gone, and we're left with
CheckProp/PopC, so we just need to dce the CheckProp.

Finally, the CheckProps are there in case a more derived class already
initialized the property; so check for AttrNoOverride on the Class,
and kill the CheckProp if its set (we could in fact analyze the
hierarchy, and check for a more-derived property of the same name -
but this gets most of the wins).

Reviewed By: ricklavoie

Differential Revision: D4689186

fbshipit-source-id: 316294cfd1530e084a3f909253ad87e5f1ab2f65