Properly handle AttrSkipFrame functions/methods
Summary:
Builtins, whether functions or methods, can be marked with AttrSkipFrame. This
means that a frame corresponding to such a function is skipped when searching
for a target frame. For example, if the skip-frame function calls
get_defined_vars(), it will return the vars of the function's caller, not its
own. Likewise, if you eval or assert, the dynamic expression will be evaluated
with the context of the skip-frame function's caller, not the skip-frame
function itself. Its main purpose is for the various wrapper functions used for
calling builtins, and for "macro-like" functions like array_map, which should
evaluate its callback as if array_map's caller had executed the callback.

Unfortunately our support for this was incomplete. The HHBC implementation of
array_map and others failed to mark themselves as skip-frame, meaning their
callback was executed in the environment of the array_map itself. We also failed
to account for the possibility of consecutive skip-frames. This is fairly
obscure, but can be done by nesting calls to array_map and others.

However, most importantly, our optimizations did not account for skip-frame
functions. Functions (like array_map) which themselves do not affect the
caller's frame can be coerced into doing so by being passed a callback which
does. Since array_map and others are builtins who aren't marked as being frame
affecting, we would assume the type of the locals would not change as a result
of the function call. If they actually do (say because array_map executes assert
with an eval string which modifies the locals), we would get incorrect results
or even crash.

Fix both HHBBC and the JIT to account for this possibility. If the function or
method is a builtin and marked as skip-frame, assume it can affect the caller's
frame. However, if calling environment affecting functions is forbidden (such as
in repo-mode), then we know skip-frame functions are safe. The reason for this
is as follows: If the builtin itself modifies the caller's frame, or called a
function which does (directly), the builtin should be marked as frame
affecting. Otherwise the only way it can affect the caller's frame is if it
calls such a function indirectly. If it can't call such functions directly, the
skip-frame function is safe.

Uncomment the existing parts of test cases which previously caused the JIT to
crash. The expanded size of the test cases causes preg_match in the test runner
to fail, so refactor how they report warnings to avoid needing wildcards. Add a
new test case full of perverse things which caused the JIT to crash previously.

Reviewed By: markw65

Differential Revision: D4678152

fbshipit-source-id: e5f0426b02375530c04e9e0b1a98e5c0be4ead9c