Tweak iva encoding
Summary:
The current format sets the LSB of the first byte to indicate
a 4-byte value in the stream. This requires the parser to right-
shift the result for both small and large values.

The new format sets the MSB of the first byte instead. This
allows the "large-value-test" to just check for a negative
value, and more importantly allows the small-value case to
return the raw value with no right-shift. This saves instructions
in the common case.

Reviewed By: markw65

Differential Revision: D4675796

fbshipit-source-id: 66b11f9045319620dc1b5da5e4a71bbc55b9fe7b