Added missing case for expliicit `and` operator in FFP lowerer
Summary:
hubyrod ran into this bug. Turned out the lowerer didn't have a case for `and` being used as an operator instead of `&&`.

hubyrod also commented in discussion that (in hphpd):
```
localhost> $a = true
localhost> $b = false
localhost> $c = $a and $b
localhost> =$c
true
```
which seems very unexpected.

For this fix, I work under the assumption that we don't want to codegen for all of PHP, with all of its quirks, but instead normalise to "nice" Hack semantics. This means that the translation from `and` to `&&` is //not// semantics preserving. That said, there is no constructor in the AST datatype to cover this scenario (because it's not in Hack).

Reviewed By: hubyrod

Differential Revision: D4697929

fbshipit-source-id: 7dee463a7978d071c700768b74d577c5e0b024e3