Associate trivia following an XHPExpression with its final token
Summary:
To ensure that trivia following an XHPOpen is added as leading trivia to the first XHPBody token rather than trailing trivia on the last XHPOpen token (since it is semantically significant for XHPBody tokens), we haven't been scanning trivia following > or /> tokens. However, trailing trivia *should* be associated with a > or /> token if the token was the last in an XHPExpression--otherwise it is associated with whatever token follows (even if it isn't the beginning of an XHPBody).

Since we only want this behavior for trailing trivia immediately after an XHPOpen's > token, this patch constrains that behavior to that one case.

Reviewed By: eshrews

Differential Revision: D4691547

fbshipit-source-id: 1749f007bfe5907692e559518eb1581d9f53173f