Allow tests to selectively enable the enforce_array_subtyping typechecker option
Summary:
This diff refactors the `GlobalOptions` a little so that tests have the opportunity to craft custom `GlobalOptions` types. I've exposed the `GlobalOptions` type in its `mli` file, because I want functions to be able to mint particular instances of the `GlobalOptions` type, which was not previously possible.

Further, I've updated `hh_single_type_check` to take advantage of this feature in order to set `enforce_array_subtyping`.

Reviewed By: dlreeves

Differential Revision: D4670053

fbshipit-source-id: 31a9af1de2bfd141173d0caad15c4c102643d79f