Codegen: command-line options for `Hack.Lang.IntsOverflowToInts` and `Eval.EnableHipHopSyntax`
Summary:
HHVM supports a number of compiler flags that influence generated code.
We now support two of these, using the same command-line syntax of
`-v Hack.Lang.IntsOverflowToInts=<value>` and `-v Eval.EnableHipHopSyntax=<value>`. When `hh_single_compile` falls back to HHVM, it passes the configuration options on.

The use of a global ref to hold the options is a bit hacky, and will be replaced by a proper environment in a future diff.

Reviewed By: hubyrod

Differential Revision: D4705705

fbshipit-source-id: b98f5425ad71a41ac3989e454430ad899c0bdbf3