Fix parsing / lexing of XHP opening tags
Summary:
Due to bugs in hphpd, I was fooled early on into believing that it was illegal to have whitespace between the < and the name of an XHP opening tag.   I therefore made this illegal in the lexer, and the parser considered the <abc  to be all one token.

This is wrong; it is legal to have whitespace in there, and the open angle and the name are two different tokens.

I've updated the parser and lexer to reflect this fact, and updated the test cases.

Differential Revision: D4660072

fbshipit-source-id: a474cd8aa119bf9e6db20c696f98b1cd951de5d5