Treat markup in PHP files as comments
Summary:
PHP supports embedding HTML tags by suspending PHP and resuming it:

    <?php
    if (foo()) {
    ?>
    <foo>bar</foo>
    <?php
    } else

We do not support this in Hack, but we would like to be able to parse PHP files in this format.

We now treat ?> and <?php  as the beginning and end of a delimited comment.

This eliminates the need for a "footer" production; in an earlier change I made ?> parse at the end of a program, again for compatibility with PHP.

At some point we'll need to produce an error if this occurs in a Hack file. That will come in a later diff.

Differential Revision: D4661098

fbshipit-source-id: 6b6a7348b7ea757cd4f1959c832343f2355a836b