Refactor memoizer to be post-processsing pass
Summary:
The function / method memoizer was running *while* compiling functions / methods.  This turns out to not be a good fit for memoizing static methods, which need to have details of the containing class, or for instance methods, which must generate a property on that class. It's just generally easier to do this after the class has already been created, and then we can do a transformation pass later.

Redoing this plumbing also greatly simplifies the code in the method and function emitters. The complexity of that code should have been a sign that this was being done too early. That complexity should be captured in the memoization rewriter, not in the function / method compilation code.

Reviewed By: andrewjkennedy

Differential Revision: D4702101

fbshipit-source-id: 6ab11412dd7ac3b5f49590eab30961ec7f616863