Make "stack" member of CallEffects more precise
Summary:
The "stack" member of CallEffects indicates which stack locations a
call may read or write from. However, all the call instructions set this to
AStkAny, which is correct, but not optimal. Change this to properly indicate the
stack locations which might actually be read or written from. A call will
generally read/write to its act-rec, its parameters, or any stack slot below the
act-rec or params.

While I'm at it, rename "destroyLocals" to "writeLocals" since destroy implies
to me killing, not writing. Plus its more symmetric with reads.

Reviewed By: markw65

Differential Revision: D4701599

fbshipit-source-id: 5541e0b259fe85478e262809e40f9c4081b3221a