Set AttrUnique and AttrPersistent in hhbbc
Summary:
This moves the logic for setting AttrUnique and AttrPersistent to
hhbbc.

It also fixes a long-standing issue with AttrPersistent: in hphpc we
would mark something persistent if it was unique, and it wasn't on a
user supplied blacklist. This lead to subtle bugs when eg autoloading
a class was supposed to trigger some side effect, and makes
repo-authoritative mode harder to use outside fb (where we pretty much
don't allow such side effects). So now we look at the pseudomain, and
if it does anything other than define classes, functions and constants
we don't allow any of its classes or functions to be AttrPersistent.

This does mean that any file used as an "entry point" (eg the start of
a script or a web page) will have all its classes and functions marked
non-persistent - but that doesn't seem like too much of a problem.

Reviewed By: jano

Differential Revision: D4689219

fbshipit-source-id: 3e478dc447ae67ed4b1e143167ee8cb8df645bc1