Refactor GC root scanning
Summary:
Before this diff, heap roots were discovered using ad-hoc code,
passing around a Scanner, and filling it with root pointers. The
heap graph "roots" consisted of these root pointers, with string
descriptions of their "seat", like CppTls, RdsLocal, and so on.

This diff refactors the ad-hoc code to pass around a lambda,
which is called with (ptr, size, type_scan::Index) for each
fine-grained root "object". Each root has a type index, allowing
it to be scanned for pointers, but also allowing us to extract
it's type name, which is nicer than the ad-hoc "seat" we had
before.

HeapGraph nodes now contain a bool indicating whether the node
is a root or not, and now all edges have valid from/to node ids,
including the pointers from root nodes. In addition, HeapGraph
now includes a list of root node ids.

ext_heapgraph was slightly extended to pass this root node
information through, but mostly left unuchanged, for compatibility
with existing clients.

Reviewed By: intellectuallyswole

Differential Revision: D4642654

fbshipit-source-id: e63abf1ad0ec5bded187a0723aa379b25a8031f5