Codegen: lambdas and function application
Summary:
Initial attempt at generating code for lambdas.
We introduce a new phase that does an AST to AST transformation at the very start of code-gen,
pulling code from lambdas out into "closure classes", and leaving a lambda with an empty body behind
that just records the "class number" necessary to generate the `CreateCl` instruction. The code generator itself then has an easy job.

Future improvements: follow HHVM more closely with respect to the list of captured variables. Generate the proper `.declvars` in the body of the closure method.

Reviewed By: hubyrod

Differential Revision: D4717328

fbshipit-source-id: 38b23f37821d2cfc78fb2c5c512175b84e161dc0