[AArch64] Adds relocation support.
Currently only static relocation is supported. Additionally
vasm jmps and jccs are optimized during relocation from their
current implementation of a ldr with embedded target literal to
a direct branch. This idion is identical to smashable jccs and jmps.
Once live relocation is fully supported, then additionally any
smashed smashable jccs and jmps can be optimized as well. This
requires yet-to-be-implemented tracking of smashable locations
and if/when they were smashed. This meta data can be reconstructed
into the CGMeta via the parsing of debug info.