Fix HipHopSyntax handling for eval and related functionality
Summary:
Eval, and other functionality which executes dynamically generated code, all
prepend <?php to their synthesized unit. If EnableHipHopSyntax is not enabled,
this means that the generated code will all be interpreted as a PHP file and
won't have HipHop specific things available to it. For example, this means that
Hack arrays cannot be used inside eval() or assert() without causing a parse
error. Instead, prepend <?hh if the file that's triggering the eval is a HH
file, and <?php otherwise.

We also weren't mangling in the value of EnableHipHopSyntax into the MD5 for the
evaled code block. This meant that if you evaled a block of code with
EnableHipHopSyntax disabled, then evaled the same code again with it enabled,
you would get the same unit as it disabled, even if the option changed behavior.

Reviewed By: markw65

Differential Revision: D4720620

fbshipit-source-id: 7fc19d9470a73d30ecc3ce46a52412d280505376